+++ src/domain/providers/composer/composerConfig.ts
@@ -20,7 +20,7 @@ export class ComposerConfig implements IProviderConfig {
     throwUndefinedOrNull(def.http, http);
   }
 
-  readonly fileLanguage = 'json';
+  readonly fileLanguage = ['json', 'jsonc'];
 
   get filePatterns(): string {
     return this.config.get(ComposerFeatures.FilePatterns);