+++ test/unit/domain/parsers/all.tests.ts
@@ -2,3 +2,4 @@ export * from './gomod/parsePackagesGoMod.tests';
 export * from './json/parsePackagesJson.tests';
 export * from './toml/parsePackagesToml.tests';
 export * from './xml/xmlDoc.tests';
+export * from './yaml/yamlUtils.tests';